+++ src/resources/resources.service.ts
